Getting everything sorted out
Have my notes to work through to finish the rest of the functionality.
Will have to make further modifications to migrations still. A navbar
boolean to a comic, as an uploaded ‘page’ should track the uploader.
And chapters need a number, in case the structure of a comic should
ever want to be Quentin-Tarantino-ized. Want to implement a draggable
gem, but gotta sort out bootstrap-sass. Worried about deployment, last
time a whole lot of asset pipeline issues came up, but the environment
should be more simple this time. When I start working on the front end
and need to test, I shouldn’t forget that the tool I need is called
‘ngrok’. Oh, chapters need a cover image; It can be a page object, I
guess, but that’s another migration modification (I don’t see the point
in piling up migrations when I don’t have any data to preserve). I need
to actually plan how I want the front to look about now, too.