Bookmarks (Cookies) p2
It doesn’t look like it, but this was a lot of work, because I
originally had methods defined near the end of the </body> inside a
<script>, and document.getElementById kept returning null. I started
navigating nodes with onclicks that I put on the submit buttons (after
I ridiculously tried wrapping the forms in a hrefs that had onclicks
before I even considered the possibility that button_tags might support
o’clock properties). I was able to hide parentNodes, but navigating to
proper childNodes was proving impossible. I finally said,
“getElementById ain’t workin’ where everyone says it should, so I
wonder… will it work inside the onclick?” By Jove, it did.